Refactor get_output_dim
